requirements.txt
asgiref==3.6.0
astroid==2.15.0
autopep8==2.0.2
dill==0.3.6
Django==4.1.7
djangorestframework==3.14.0
isort==5.12.0
jedi==0.18.2
lazy-object-proxy==1.9.0
mccabe==0.7.0
parso==0.8.3
platformdirs==3.1.1
pycodestyle==2.10.0
pylint==2.17.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
wrapt==1.15.0
